# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: BTP User Containment API
  version: 0.1.0-SNAPSHOT
  description: |
    REST API for the OOBS: a central operator console
    that enrolls SAP BTP subaccounts / IAS tenants, discovers identity providers
    and SoD conflicts, and contains (locks) or unlocks compromised users across subaccounts,
    Cloud Foundry (XSUAA) and IAS in response to an incident.

    ## Authentication
    All `/api/v1/**` endpoints (except `/health` and `/whoami`) require a valid
    XSUAA-issued OAuth2 **bearer JWT**. Authorization is enforced per-operation
    via OAuth2 scopes; the required scope is noted in each operation description.

    ## CSRF
    Keep in mind that the SAP app router uses CSRF. 
    For that see: x-csrf-token docu on the sap help site 
    (it changes all x days so you need to search it)

    ## Scope model
    Four scopes exist, in increasing privilege. The `@PreAuthorize` rules map to:
    - **`btpc.viewer`** - read-only. Satisfied by any of the four scopes.
    - **`btpc.sod_viewer`** - may run SoD scans (admin also allowed).
    - **`btpc.responder`** - may contain/unlock users (admin also allowed).
    - **`btpc.admin`** - full control: enrollment, credentials, config, deletes.

    "Required scope" below names the *minimum* intent; `btpc.admin` always
    satisfies it.

    ## Caller source
    Mutating endpoints accept an optional `X-Caller-Source` header used only for
...
    not change behaviour or authorization.
  license:
...
  contact:
...
servers:
...
security:
  - xsuaa: []
paths:
...
  /api/v1/subaccounts:
    get:
      tags: [Subaccounts]
      summary: List enrolled subaccounts
      description: 'Required scope: `btpc.viewer`.'
      responses:
        '200':
          description: Enrolled subaccounts
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SubaccountDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Subaccounts]
      summary: Enroll a subaccount from a service key
      description: 'Required scope: `btpc.admin`.'
      parameters:
        - $ref: '#/components/parameters/CallerSource'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EnrollRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubaccountDto' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/subaccounts/quick-add:
    post:
      tags: [Subaccounts]
      summary: Manually add a subaccount (no central-viewer required)
      description: 'Required scope: `btpc.admin`.'
      parameters:
        - $ref: '#/components/parameters/CallerSource'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/QuickAddRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubaccountDto' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/subaccounts/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Subaccounts]
      summary: Get one subaccount
      description: 'Required scope: `btpc.viewer`.'
      responses:
        '200':
          description: Subaccount
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubaccountDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Subaccounts]
      summary: Unenroll a subaccount
      description: 'Required scope: `btpc.admin`.'
      parameters:
        - $ref: '#/components/parameters/CallerSource'
      responses:
        '204': { description: Unenrolled }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/subaccounts/{id}/label:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    patch:
      tags: [Subaccounts]
      summary: Update the display label
      description: 'Required scope: `btpc.admin`.'
      parameters:
        - $ref: '#/components/parameters/CallerSource'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LabelPatch' }
      responses:
        '200':
          description: Updated subaccount
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubaccountDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/subaccounts/{id}/metadata:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    patch:
      tags: [Subaccounts]
      summary: Update subaccount metadata
      description: 'Required scope: `btpc.admin`.'
      parameters:
        - $ref: '#/components/parameters/CallerSource'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MetadataPatch' }
      responses:
        '200':
          description: Updated subaccount
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubaccountDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/subaccounts/{id}/credentials:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [Subaccounts]
      summary: Attach a service key credential
      description: 'Required scope: `btpc.admin`. `serviceKey` is required.'
      parameters:
        - $ref: '#/components/parameters/CallerSource'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AttachCredential' }
      responses:
        '200':
          description: Updated subaccount
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubaccountDto' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/subaccounts/{id}/cf-org-id:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    patch:
      tags: [Subaccounts]
      summary: Set the Cloud Foundry org id
      description: |
        Required scope: `btpc.admin`. Partial update of a single subaccount
        attribute (consistent with the `label` and `metadata` patches). Send
        `{ "cfOrgId": null }` to clear the pin.
      parameters:
        - $ref: '#/components/parameters/CallerSource'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetCfOrgId' }
      responses:
        '200':
          description: Updated subaccount
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubaccountDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/subaccounts/{id}/cf-technical-user:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [Subaccounts]
      summary: Attach a CF technical user
      description: 'Required scope: `btpc.admin`.'
      parameters:
        - $ref: '#/components/parameters/CallerSource'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CfTechnicalUserPayload' }
      responses:
        '200':
          description: Updated subaccount
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubaccountDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/subaccounts/{id}/cf-technical-user/copy-from/{sourceId}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
      - name: sourceId
        in: path
        required: true
        description: Subaccount to copy the technical-user credential from
        schema: { type: string, format: uuid }
    post:
      tags: [Subaccounts]
      summary: Copy a CF technical user from another subaccount
      description: 'Required scope: `btpc.admin`.'
      parameters:
        - $ref: '#/components/parameters/CallerSource'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CopyCfTechUser' }
      responses:
        '200':
          description: Updated subaccount
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubaccountDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/subaccounts/{id}/provision-xsuaa-apiaccess:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [Subaccounts]
      summary: Provision an XSUAA apiaccess plan instance
      description: 'Required scope: `btpc.admin`.'
      parameters:
        - $ref: '#/components/parameters/CallerSource'
      responses:
        '200':
          description: Updated subaccount
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SubaccountDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ------------------------------------------------------------------ Contacts
...
  /api/v1/ias-tenants:
    get:
      tags: [IAS tenants]
      summary: List enrolled IAS tenants
      description: 'Required scope: `btpc.viewer`.'
      responses:
        '200':
          description: IAS tenants
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/IasTenantDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [IAS tenants]
      summary: Enroll an IAS tenant
      description: 'Required scope: `btpc.admin`.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IasTenantCreatePayload' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IasTenantDto' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/v1/ias-tenants/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [IAS tenants]
      summary: Get one IAS tenant
      description: 'Required scope: `btpc.viewer`.'
      responses:
        '200':
          description: IAS tenant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IasTenantDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [IAS tenants]
      summary: Delete an IAS tenant
      description: 'Required scope: `btpc.admin`.'
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/ias-tenants/{id}/meta:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    patch:
      tags: [IAS tenants]
      summary: Update IAS tenant display name
      description: 'Required scope: `btpc.admin`.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IasTenantMeta' }
      responses:
        '200':
          description: Updated tenant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IasTenantDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/ias-tenants/{id}/credentials:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    put:
      tags: [IAS tenants]
      summary: Replace IAS tenant credentials
      description: 'Required scope: `btpc.admin`.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IasTenantCreatePayload' }
      responses:
        '200':
          description: Updated tenant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IasTenantDto' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ------------------------------------------------------------ Origin profiles
...
components:
  securitySchemes:
    xsuaa:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        XSUAA-issued OAuth2 access token. Scopes (`btpc.admin`,
        `btpc.responder`, `btpc.sod_viewer`, `btpc.viewer`) are carried as
        `SCOPE_<name>` authorities and enforced per operation.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    CallerSource:
      name: X-Caller-Source
      in: header
      required: false
      description: Audit attribution only; default UI.
      schema:
        type: string
        enum: [soar, soar-api, system]
  responses:
    BadRequest:
      description: Invalid input or malformed JSON body
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid bearer token
    Forbidden:
      description: Authenticated but missing the required scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Conflict:
      description: State conflict or data-integrity violation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      description: Uniform error envelope. `null` fields are omitted.
      properties:
        error:
          type: string
          description: Stable machine code.
          enum: [bad_request, malformed_body, not_found, conflict, cis_upstream, forbidden, server_error]
        message:
          type: string
          nullable: true
        timestamp:
          type: string
          format: date-time

    # ---- Platform ----
...
    LabelPatch:
      type: object
      properties:
        label: { type: string, nullable: true }
    AttachCredential:
      type: object
      required: [serviceKey]
      properties:
        serviceKey:
          type: string
          description: Raw BTP service-key JSON (string).
    SetCfOrgId:
      type: object
      properties:
        cfOrgId: { type: string, format: uuid }
    CopyCfTechUser:
      type: object
      properties:
        cfApiUrl: { type: string }
        cfUaaUrl: { type: string }
    Stage:
      type: string
      description: Deployment stage. Validated case-insensitively against this set.
      enum: [prod, dev, qa, test]
      nullable: true
    EnrollRequest:
      type: object
      description: >
        SubaccountService.EnrollRequest. Carries the subaccount identity and the
        service keys to wrap and store. Pasted IAS keys are ignored (manage IAS at
        the tenant level instead).
      required: [subaccountGuid]
      properties:
        subaccountGuid:
          type: string
          format: uuid
          description: Required. Pick it from the Discover picker or type it manually.
        cisDisplayName: { type: string, nullable: true, description: 'Falls back to the CIS key identity zone, else "Subaccount <guid>".' }
        region: { type: string, nullable: true, description: 'Falls back to the CIS key region, else "unknown".' }
        globalAccountId: { type: string, format: uuid, nullable: true }
        globalAccountName: { type: string, nullable: true }
        stage: { $ref: '#/components/schemas/Stage' }
        label: { type: string, nullable: true }
        serviceKeys:
          type: array
          description: Raw BTP service-key JSON documents (each a string). Classified on enroll.
          items: { type: string }
        discoveredId:
          type: string
          format: uuid
          nullable: true
          description: Discovery candidate this enrollment was promoted from (audit only).
    QuickAddRequest:
      type: object
      description: Manually register a subaccount with no central-viewer key. No credential is attached.
      required: [subaccountGuid, cisDisplayName, region]
      properties:
        subaccountGuid: { type: string, format: uuid }
        cisDisplayName: { type: string, description: Display name (the "name" field in the UI). }
        region: { type: string }
        globalAccountId: { type: string, format: uuid, nullable: true }
        globalAccountName: { type: string, nullable: true }
        stage: { $ref: '#/components/schemas/Stage' }
        label: { type: string, nullable: true }
    MetadataPatch:
      type: object
      description: Partial update of subaccount metadata. Omitted/blank fields are cleared to null.
      properties:
        label: { type: string, nullable: true }
        globalAccountId: { type: string, format: uuid, nullable: true }
        globalAccountName: { type: string, nullable: true }
        stage: { $ref: '#/components/schemas/Stage' }
    CfTechnicalUserPayload:
      type: object
      description: >
        CF technical-user credential (IAS X.509 passcode flow). All fields are
        required; the server runs a live `/v3/organizations` smoke test before
        storing and rejects the credential if it fails.
      required: [cfApiUrl, cfUaaUrl, username, origin, iasPasscodeUrl, p12Base64, p12Password]
      properties:
        cfApiUrl: { type: string, description: Cloud Controller API URL (allowlist-checked). }
        cfUaaUrl: { type: string, description: UAA token URL (allowlist-checked). }
        username: { type: string }
        origin: { type: string, description: 'IAS origin key - what `cf login --origin <key>` uses (Trust Config → *-platform).' }
        iasPasscodeUrl: { type: string, description: IAS passcode endpoint (allowlist-checked). }
        p12Base64: { type: string, description: Base64-encoded PKCS#12 client cert. }
        p12Password: { type: string }
    SubaccountDto:
      type: object
      description: >
        Enrolled subaccount projection (ninja.gruber.btpc.enroll.domain.SubaccountDto).
        Null fields are omitted from the response.
      properties:
        id: { type: string, format: uuid }
        subaccountGuid: { type: string, format: uuid }
        cisDisplayName: { type: string }
        cisDisplayNameRefreshedAt: { type: string, format: date-time, nullable: true }
        label: { type: string, nullable: true }
        region: { type: string }
        globalAccountId: { type: string, format: uuid, nullable: true }
        globalAccountName: { type: string, nullable: true }
        stage: { type: string, nullable: true }
        enrolledBy: { type: string }
        enrolledAt: { type: string, format: date-time }
        lastHealthAt: { type: string, format: date-time, nullable: true }
        lastHealthError: { type: string, nullable: true }
        status: { type: string, enum: [active, disabled, error] }
        cfOrgId: { type: string, format: uuid, nullable: true }
        capabilities: { $ref: '#/components/schemas/SubaccountCapabilities' }
        contactCount: { type: integer }
    SubaccountCapabilities:
      type: object
      description: Which credential kinds are attached to the subaccount.
      properties:
        cis: { type: boolean }
        xsuaaApiaccess: { type: boolean }
        cfTechnicalUser: { type: boolean }

    # ---- Contacts ----
...
    IasTenantCreatePayload:
      type: object
      description: >
        IasTenantService.CreatePayload. Also used to replace credentials (PUT
        .../credentials), in which case the URL host must match the existing tenant.
      required: [displayName, url, p12Base64]
      properties:
        displayName: { type: string }
        url: { type: string, description: IAS tenant URL; the host is derived and used as the unique key (allowlist-checked). }
        p12Base64: { type: string, description: Base64-encoded PKCS#12 from the X509_GENERATED IAS application binding. }
        p12Password: { type: string, nullable: true, description: Defaults to empty if omitted. }
    IasTenantMeta:
      type: object
      properties:
        displayName: { type: string }
    IasTenantDto:
      type: object
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string, nullable: true }
        iasHost: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        createdBy: { type: string, nullable: true }
        updatedAt: { type: string, format: date-time, nullable: true }
        updatedBy: { type: string, nullable: true }

    # ---- Origin profiles / IdP discovery ----
...